new version, almost complete rewrite:  - only 1 julia process (not one per editor)  - linting dynamically  - feedback on start of environment parsing  - faster startup  - TODO: shutdown on windows
