+++ lib/index.js
@@ -1,11 +1,12 @@
 /* @flow */
 
 import net from 'net';
+import FS from 'fs';
 import { CompositeDisposable } from 'atom';
-import { spawnServer, terminateServer } from './server';
-import type { Server } from './types';
+import { spawnServer, terminateServer, getPipePath } from './server';
 
-let spawnedServer: ?Server = null;
+const pipepath = getPipePath();
+
 let subscriptions: ?Object = null;
 
 let executablePath;
@@ -14,6 +15,26 @@ let ignoreWarning;
 let showErrorCodes;
 let ignoreIssueCodes;
 
+function timeout(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+async function pipeup() {
+  if (!FS.existsSync(pipepath)) {
+    await timeout(1000);
+    await pipeup();
+  }
+}
+
+async function unlock() {
+  if (global.linter_julia_locked === true) {
+    await timeout(200);
+    await unlock();
+    return;
+  }
+  global.linter_julia_locked = true;
+}
+
 export function activate() {
   // eslint-disable-next-line global-require
   require('atom-package-deps').install('linter-julia');
@@ -21,11 +42,12 @@ export function activate() {
   subscriptions.add(
     atom.config.observe('linter-julia.executablePath', async (value) => {
       executablePath = value;
-      if (spawnedServer) {
+      if (global.linter_julia_spawnedServer) {
         try {
-          await terminateServer(spawnedServer);
-          spawnedServer = null;
-          spawnedServer = await spawnServer(executablePath);
+          await terminateServer(global.linter_julia_spawnedServer);
+          global.linter_julia_spawnedServer = spawnServer(executablePath);
+          global.linter_julia_started = true;
+          global.linter_julia_locked = false;
         } catch (e) {
           const message = '[Linter-Julia] '
             + 'Unable to spawn server after config change';
@@ -48,12 +70,21 @@ export function activate() {
       ignoreIssueCodes = value;
     }),
   );
+
+  // start only one server
+  if (typeof global.linter_julia_started === 'undefined') {
+    global.linter_julia_spawnedServer = spawnServer(executablePath);
+    global.linter_julia_started = true;
+    global.linter_julia_locked = false;
+  }
 }
 
 export function deactivate() {
-  if (spawnedServer) {
-    terminateServer(spawnedServer);
-    spawnedServer = null;
+  if (global.linter_julia_spawnedServer !== null) {
+    terminateServer(global.linter_julia_spawnedServer);
+    global.linter_julia_spawnedServer = null;
+    global.linter_julia_started = false;
+    global.linter_julia_locked = false;
   }
   if (subscriptions) {
     subscriptions.dispose();
@@ -67,10 +98,14 @@ export function provideLinter() {
     lintsOnChange: true,
     grammarScopes: ['source.julia'],
     async lint(textEditor: Object) {
-      if (!spawnedServer) {
-        spawnedServer = await spawnServer(executablePath);
-      }
-      const connection = net.createConnection(spawnedServer.path);
+      // wait for the pipe to appear
+      await pipeup();
+
+      // only one connection at one time
+      await unlock();
+
+      const connection = net.createConnection(pipepath);
+
       connection.on('connect', function writeData() {
         this.write(JSON.stringify({
           file: textEditor.getPath(),
@@ -107,6 +142,14 @@ export function provideLinter() {
             resolve(null);
             return;
           }
+          if (parsed.length > 0) {
+            if (parsed[0].description === 'I000') {
+              atom.notifications.addInfo(`linter-julia: please wait - generating symbols for environment ${parsed[0].excerpt}`);
+              parsed = [];
+            }
+          }
+
+          global.linter_julia_locked = false;
           resolve(parsed);
         });
       }));
