+++ lib/server.js
@@ -3,17 +3,15 @@
 import os from 'os';
 import FS from 'fs';
 import Path from 'path';
-import uuid from 'uuid4';
 import { async as getAsyncEnv } from 'consistent-env';
 import { BufferedProcess, Disposable, CompositeDisposable } from 'atom';
 import type { Server } from './types';
 
 const JULIA_SERVER_PATH = Path.join(__dirname, 'julia-server.jl');
 
-export async function getPipePath(): Promise<string> {
+export function getPipePath() {
   const baseDir = process.platform === 'win32' ? '\\\\.\\pipe\\' : `${os.tmpdir()}/`;
-  const uniqueId = uuid();
-  return baseDir + uniqueId;
+  return `${baseDir}linter_julia_${process.pid}`;
 }
 
 export async function spawnServer(juliaExecutable: string): Promise<Server> {
